requirement.txt
python-dateutil
requests
plyer
schedule